requirements.txt
Flask==3.1.0
gunicorn==23.0.0
mysql-connector-python==9.3.0
python-dotenv==1.1.1
requests==2.32.3
reportlab==4.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.66.5
